renamed project middle name
