+++ Sauti/app/src/main/java/com/labs/sauti/di/component/LoginComponent.kt
@@ -1,0 +1,12 @@
+package com.labs.sauti.di.component
+
+import com.labs.sauti.activity.LoginActivity
+import com.labs.sauti.di.module.LoginModule
+import dagger.Subcomponent
+
+@Subcomponent(modules = [LoginModule::class])
+interface LoginComponent {
+
+    fun inject(loginActivity: LoginActivity)
+
+}
