+++ Sauti/app/src/main/java/com/labs/sauti/di/component/UserComponent.kt
@@ -1,0 +1,10 @@
+package com.labs.sauti.di.component
+
+import com.labs.sauti.activity.SearchActivity
+import com.labs.sauti.di.module.UserModule
+import dagger.Subcomponent
+
+@Subcomponent(modules = [UserModule::class])
+interface UserComponent {
+    fun inject(searchActivity: SearchActivity)
+}
