+++ Sauti/app/src/main/java/com/labs/sauti/di/module/AppModule.kt
@@ -1,0 +1,17 @@
+package com.labs.sauti.di.module
+
+import android.content.Context
+import dagger.Module
+import dagger.Provides
+import javax.inject.Singleton
+
+@Module
+class AppModule(private val context: Context) {
+
+    @Provides
+    @Singleton
+    fun provideAppContext(): Context {
+        return context.applicationContext
+    }
+
+}
