+++ Sauti/app/src/main/java/com/labs/sauti/di/module/LoginModule.kt
@@ -1,0 +1,16 @@
+package com.labs.sauti.di.module
+
+import com.labs.sauti.repository.SautiRepository
+import com.labs.sauti.view_model.LoginViewModel
+import dagger.Module
+import dagger.Provides
+
+@Module
+class LoginModule {
+
+    @Provides
+    fun provideLoginViewModelFactory(sautiRepository: SautiRepository): LoginViewModel.Factory {
+        return LoginViewModel.Factory(sautiRepository)
+    }
+
+}
