+++ Sauti/app/src/main/java/com/labs/sauti/model/SautiApiError.kt
@@ -1,0 +1,3 @@
+package com.labs.sauti.model
+
+class SautiApiError(var message: String)
