+++ Sauti/app/src/main/java/com/labs/sauti/model/User.kt
@@ -1,0 +1,14 @@
+package com.labs.sauti.model
+
+import com.google.gson.annotations.Expose
+import com.google.gson.annotations.SerializedName
+
+class User {
+    @SerializedName("userid")
+    @Expose
+    var id: Long? = null
+
+    @SerializedName("username")
+    @Expose
+    var username: String? = null
+}
