+++ Sauti/app/src/main/java/com/labs/sauti/repository/SautiRepository.kt
@@ -1,0 +1,12 @@
+package com.labs.sauti.repository
+
+import com.labs.sauti.model.LoginResponse
+import com.labs.sauti.model.User
+import io.reactivex.Single
+
+interface SautiRepository {
+
+    fun login(username: String, password: String): Single<LoginResponse>
+    fun getCurrentUser(accessToken: String): Single<User>
+
+}
