+++ Sauti/app/src/main/java/com/labs/sauti/view_model/BaseViewModel.kt
@@ -1,0 +1,23 @@
+package com.labs.sauti.view_model
+
+import androidx.lifecycle.ViewModel
+import io.reactivex.disposables.CompositeDisposable
+import io.reactivex.disposables.Disposable
+
+open class BaseViewModel : ViewModel() {
+
+    private val compositeDisposable: CompositeDisposable = CompositeDisposable()
+
+    protected fun addDisposable(disposable: Disposable) {
+        compositeDisposable.add(disposable)
+    }
+
+    private fun clearDisposables() {
+        compositeDisposable.clear()
+    }
+
+    override fun onCleared() {
+        clearDisposables()
+    }
+
+}
