--- Sauti/app/src/main/java/com/sauti/sauti/di/component/MainComponent.kt
@@ -1,12 +1,0 @@
-package com.sauti.sauti.di.component
-
-import com.sauti.sauti.di.module.*
-import dagger.Component
-import javax.inject.Singleton
-
-@Singleton
-@Component(modules = [AppModule::class, NetworkModule::class, DataModule::class])
-interface MainComponent {
-    fun plus(loginModule: LoginModule): LoginComponent
-    fun plus(userModule: UserModule): UserComponent
-}
