--- Sauti/app/src/main/java/com/sauti/sauti/di/module/AppModule.kt
@@ -1,17 +1,0 @@
-package com.sauti.sauti.di.module
-
-import android.content.Context
-import dagger.Module
-import dagger.Provides
-import javax.inject.Singleton
-
-@Module
-class AppModule(private val context: Context) {
-
-    @Provides
-    @Singleton
-    fun provideAppContext(): Context {
-        return context.applicationContext
-    }
-
-}
