--- Sauti/app/src/main/java/com/sauti/sauti/model/SautiApiError.kt
@@ -1,3 +1,0 @@
-package com.sauti.sauti.model
-
-class SautiApiError(var message: String)
