--- Sauti/app/src/main/java/com/sauti/sauti/model/User.kt
@@ -1,14 +1,0 @@
-package com.sauti.sauti.model
-
-import com.google.gson.annotations.Expose
-import com.google.gson.annotations.SerializedName
-
-class User {
-    @SerializedName("userid")
-    @Expose
-    var id: Long? = null
-
-    @SerializedName("username")
-    @Expose
-    var username: String? = null
-}
