--- Sauti/app/src/main/java/com/sauti/sauti/repository/SautiRepositoryImpl.kt
@@ -1,16 +1,0 @@
-package com.sauti.sauti.repository
-
-import com.sauti.sauti.api.SautiApiService
-import com.sauti.sauti.model.LoginResponse
-import com.sauti.sauti.model.User
-import io.reactivex.Single
-
-class SautiRepositoryImpl(private val sautiApiService: SautiApiService, private val sautiAuthorization: String) : SautiRepository {
-    override fun login(username: String, password: String): Single<LoginResponse> {
-        return sautiApiService.login(sautiAuthorization, "password", username, password)
-    }
-
-    override fun getCurrentUser(accessToken: String): Single<User> {
-        return sautiApiService.getCurrentUser("Bearer $accessToken")
-    }
-}
