--- Sauti/app/src/main/java/com/sauti/sauti/view_model/UserViewModel.kt
@@ -1,43 +1,0 @@
-package com.sauti.sauti.view_model
-
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.ViewModelProvider
-import com.sauti.sauti.model.SautiApiError
-import com.sauti.sauti.model.User
-import com.sauti.sauti.repository.SautiRepository
-import com.sauti.sauti.sp.SessionSp
-
-class UserViewModel(private val sautiRepository: SautiRepository, private val sessionSp: SessionSp): BaseViewModel() {
-
-    private val userLiveData = MutableLiveData<User>()
-    private val errorLiveData = MutableLiveData<SautiApiError>()
-
-    fun getUserLiveData(): LiveData<User> = userLiveData
-    fun getErrorLiveData(): LiveData<SautiApiError> = errorLiveData
-
-    fun getCurrentUser() {
-        if (sessionSp.isAccessTokenValid()) {
-            val disposable = sautiRepository.getCurrentUser(sessionSp.getAccessToken()).subscribe(
-                {
-                    userLiveData.postValue(it)
-                },
-                {
-                    val error = SautiApiError(it.message ?: "")
-                    errorLiveData.postValue(error)
-                }
-            )
-
-            addDisposable(disposable)
-        }
-    }
-
-    class Factory(private val sautiRepository: SautiRepository, private val sessionSp: SessionSp) : ViewModelProvider.Factory {
-        override fun <T : ViewModel?> create(modelClass: Class<T>): T {
-            @Suppress("UNCHECKED_CAST")
-            return UserViewModel(sautiRepository, sessionSp) as T
-        }
-    }
-
-}
